refactor(core): introduce S2SpringAdapterBase to centralize common logic
Create a new abstract class `S2SpringAdapterBase` to encapsulate common
logic and functionality shared across different Spring adapters. This
includes methods for creating an `AutomateContext`, `GuardVerifier`,
and `AutomateEventPublisher`, as well as defining guards.

Refactor `S2AutomateDeciderSpringAdapter` and `S2ConfigurerAdapter` to
extend `S2SpringAdapterBase`, removing duplicated code and improving
maintainability. This change reduces redundancy and centralizes the
shared logic, making it easier to manage and update in the future.
