+++ s2-spring/sourcing/s2-spring-boot-starter-sourcing/src/main/kotlin/s2/spring/automate/sourcing/S2AutomateDeciderSpringAdapter.kt
@@ -7,12 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.ApplicationContext
 import org.springframework.context.ApplicationContextAware
 import org.springframework.context.support.GenericApplicationContext
-import s2.automate.core.guard.TransitionStateGuard
-import s2.automate.core.appevent.publisher.AutomateEventPublisher
-import s2.automate.core.context.AutomateContext
 import s2.automate.core.engine.S2AutomateEngineImpl
-import s2.automate.core.guard.Guard
-import s2.automate.core.guard.GuardVerifierImpl
 import s2.dsl.automate.Evt
 import s2.dsl.automate.S2Automate
 import s2.dsl.automate.S2State
@@ -27,13 +22,14 @@ import s2.sourcing.dsl.view.ViewLoader
 import s2.automate.core.storing.snap.RetryTaskChannel
 import s2.automate.core.storing.snap.SnapPersister
 import s2.spring.automate.sourcing.persist.S2AutomateSourcingPersister
+import s2.spring.core.S2SpringAdapterBase
 import s2.spring.core.publisher.SpringEventPublisher
 
 abstract class S2AutomateDeciderSpringAdapter<ENTITY, STATE, EVENT, ID, EXECUTOR>(
 	val executor: EXECUTOR,
 	val view: View<EVENT, ENTITY>,
 	val snapRepository: SnapRepository<ENTITY, ID>? = null,
-): InitializingBean, ApplicationContextAware where
+): InitializingBean, ApplicationContextAware, S2SpringAdapterBase<ENTITY, STATE, EVENT, ID>() where
 STATE : S2State,
 ENTITY : WithS2State<STATE>,
 ENTITY : WithS2Id<ID>,
@@ -79,27 +75,7 @@ EXECUTOR : S2AutomateDeciderSpring<ENTITY, STATE, EVENT, ID> {
 		}
 	}
 
-	protected open fun automateContext() = AutomateContext(automate())
-
-	protected open fun guardExecutor(
-		automateAppEventPublisher: AutomateEventPublisher<STATE, ID, ENTITY, S2Automate>,
-	): GuardVerifierImpl<STATE, ID, ENTITY, EVENT, S2Automate> {
-		return GuardVerifierImpl(
-			guards = guards(),
-			publisher = automateAppEventPublisher
-		)
-	}
-
-	protected open fun automateAppEventPublisher(eventPublisher: SpringEventPublisher)
-			: AutomateEventPublisher<STATE, ID, ENTITY, S2Automate> {
-		return AutomateEventPublisher(eventPublisher)
-	}
-
-	protected open fun guards(): List<Guard<STATE, ID, ENTITY, EVENT, S2Automate>> = listOf(
-		TransitionStateGuard()
-	)
-
-	abstract fun automate(): S2Automate
+	abstract override fun automate(): S2Automate
 	abstract fun eventStore(): EventRepository<EVENT, ID>
 	abstract fun entityType(): KClass<EVENT>
 	open fun preventOptimisticLocking(): Boolean = false
