+++ s2-spring/storing/s2-spring-boot-starter-storing/src/main/kotlin/s2/spring/automate/S2ConfigurerAdapter.kt
@@ -6,7 +6,9 @@ import s2.automate.core.engine.S2AutomateEngineImpl
 import s2.automate.core.guard.TransitionStateGuard
 import s2.automate.core.appevent.publisher.AutomateEventPublisher
 import s2.automate.core.context.AutomateContext
+import s2.automate.core.guard.Guard
 import s2.automate.core.guard.GuardAdapter
+import s2.automate.core.guard.GuardVerifier
 import s2.automate.core.guard.GuardVerifierImpl
 import s2.automate.core.persist.AutomatePersister
 import s2.dsl.automate.Evt
@@ -15,9 +17,11 @@ import s2.dsl.automate.S2State
 import s2.dsl.automate.model.WithS2Id
 import s2.dsl.automate.model.WithS2State
 import s2.spring.automate.executor.S2AutomateExecutorSpring
+import s2.spring.core.S2SpringAdapterBase
 import s2.spring.core.publisher.SpringEventPublisher
 
-abstract class S2ConfigurerAdapter<STATE, ID, ENTITY, out EXECUTER>: InitializingBean where
+abstract class S2ConfigurerAdapter<STATE, ID, ENTITY, out EXECUTER>
+	: InitializingBean, S2SpringAdapterBase<ENTITY, STATE, Evt, ID>() where
 STATE : S2State,
 ENTITY : WithS2State<STATE>,
 ENTITY : WithS2Id<ID>,
@@ -34,32 +38,13 @@ EXECUTER : S2AutomateExecutorSpring<STATE, ID, ENTITY> {
 		return S2AutomateEngineImpl(automateContext, guardExecutor, persister, publisher)
 	}
 
-	protected open fun automateContext() = AutomateContext(automate())
-
-	protected open fun guardExecutor(
-		automateAppEventPublisher: AutomateEventPublisher<STATE, ID, ENTITY, S2Automate>,
-	): GuardVerifierImpl<STATE, ID, ENTITY, Evt, S2Automate> {
-		return GuardVerifierImpl(
-			guards = guards(),
-			publisher = automateAppEventPublisher
-		)
-	}
-
-	protected open fun automateAppEventPublisher(eventPublisher: SpringEventPublisher)
-			: AutomateEventPublisher<STATE, ID, ENTITY, S2Automate> {
-		return AutomateEventPublisher(eventPublisher)
-	}
-
-	protected open fun guards(): List<GuardAdapter<STATE, ID, ENTITY, Evt, S2Automate>> =
-		listOf(TransitionStateGuard())
-
 	override fun afterPropertiesSet() {
 		val automateExecutorFlow = aggregateFlow()
 		val agg = executor()
 		agg.withContext(automateExecutorFlow, eventPublisher)
 	}
 
+	abstract override fun automate(): S2Automate
 	abstract fun aggregateRepository(): AutomatePersister<STATE, ID, ENTITY, Evt, S2Automate>
-	abstract fun automate(): S2Automate
 	abstract fun executor(): EXECUTER
 }
